+++ next.config.mjs
@@ -1,6 +1,3 @@
-const nextConfig = {
+const nextConfig = {};
 
-  };
-  
-  export default nextConfig;
-  
+export default nextConfig;
